mypractice/tool_compilations/requirements.txt
amqp==2.4.2
aniso8601==6.0.0
apscheduler==3.6.0
bcrypt==3.1.6
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
click==7.0
dnspython==1.16.0
email-validator==1.0.4
flask-apscheduler==1.11.0
flask-bcrypt==0.7.1
flask-login==0.4.1
flask-restful==0.3.6
flask-zipkin==0.0.4
flask==1.0.3
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
kombu==4.5.0
markupsafe==1.1.1
mongoengine==0.17.0
numpy==1.16.3
pandas==0.24.2
pillow==6.0.0
ply==3.11
psycopg2-binary==2.8.2
py-zipkin==0.18.3
pycparser==2.19
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.2
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
sqlalchemy==1.3.3
terminaltables==3.1.0
thriftpy2==0.4.2
tzlocal==1.5.1
urllib3==1.25.3
uwsgi==2.0.18
validr==1.1.0
vine==1.3.0
werkzeug==0.15.4
xlrd==1.2.0
xlsxwriter==1.1.8

